nirmatee_prj/tally_process/requirements.txt
django
djangorestframework
python-decouple
